Extract plotting helpers into dedicated module
